Edge-Thickness, Flux loading
Edge thickness ist jetzt proportional zum Betrag des Fluxes. Flux-file kann aber muss nicht eingeladen werden.
Zusätzlicher 'name' column im EdgeTable damit man die Edges den Fluxes genau zuordnen kann.
Code bisschen bereinigt und eine komische Stelle gefunden!!

CreateNodes Zeile 104.

+++ sample-create-network-view/src/main/java/org/cytoscape/sample/internal/CreateNetworkViewTaskFactory.java
@@ -10,7 +10,6 @@ import org.cytoscape.view.model.CyNetworkViewManager;
 import org.cytoscape.work.AbstractTaskFactory;
 import org.cytoscape.work.TaskIterator;
 
-import javax.swing.*;
 import java.util.HashMap;
 
 public class CreateNetworkViewTaskFactory extends AbstractTaskFactory {
@@ -37,7 +36,7 @@ public class CreateNetworkViewTaskFactory extends AbstractTaskFactory {
 
 	public TaskIterator createTaskIterator(){
 		FileChoosing newChooser = new FileChoosing();
-		HashMap<String, Float> csvMap = newChooser.makeMap();
+		HashMap<String, Double> csvMap = newChooser.makeMap();
 		return new TaskIterator(new CreateNetworkViewTask(cyNetworkNaming, cnf,networkManager, cnvf, networkViewManager, dataSourceManager, currentNetwork, csvMap));
 	}
 }
